Summon implicit instance for digest insteal of a direct call
