+++ sim/src/main/scala/sim/NetworkSim.scala
@@ -1,10 +1,10 @@
 package sim
 
+import cats.Parallel
 import cats.effect.*
 import cats.effect.kernel.{Async, Temporal}
 import cats.effect.std.{Console, Random}
 import cats.syntax.all.*
-import cats.{Eval, Parallel}
 import diagnostics.KamonContextStore
 import diagnostics.metrics.{Config as InfluxDbConfig, InfluxDbBatchedMetrics}
 import dproc.data.Block
@@ -20,8 +20,6 @@ import pureconfig.generic.ProductHint
 import sdk.api
 import sdk.api.data.{Bond, Deploy, Status}
 import sdk.api.{data, ExternalApi}
-import sdk.codecs.Digest
-import sdk.codecs.protobuf.ProtoPrimitiveWriter
 import sdk.diag.{Metrics, SystemReporter}
 import sdk.history.ByteArray32
 import sdk.history.History.EmptyRootHash
@@ -33,6 +31,7 @@ import sdk.syntax.all.*
 import sim.Config as SimConfig
 import sim.NetworkSnapshot.{reportSnapshot, NodeSnapshot}
 import sim.balances.*
+import sim.balances.Hashing.*
 import sim.balances.MergeLogicForPayments.mergeRejectNegativeOverflow
 import sim.balances.data.BalancesState.Default
 import sim.balances.data.{BalancesDeploy, BalancesDeployBody, BalancesState}
@@ -45,22 +44,6 @@ import scala.concurrent.duration.{Duration, DurationInt}
 object NetworkSim extends IOApp {
 
   implicit def blake2b256Hash(x: Array[Byte]): ByteArray32 = ByteArray32.convert(Blake2b.hash256(x)).getUnsafe
-
-  implicit val balancesDeployBodyDigest: Digest[BalancesDeployBody] =
-    new sdk.codecs.Digest[BalancesDeployBody] {
-      override def digest(x: BalancesDeployBody): ByteArray = {
-        val bytes = ProtoPrimitiveWriter.encodeWith(Serialization.balancesDeployBodySerialize[Eval].write(x))
-        ByteArray(Blake2b.hash256(bytes.value))
-      }
-    }
-
-  implicit val blockBodyDigest: Digest[Block[M, S, T]] =
-    new sdk.codecs.Digest[Block[M, S, T]] {
-      override def digest(x: Block[M, S, T]): ByteArray = {
-        val bytes = ProtoPrimitiveWriter.encodeWith(Serialization.blockSerialize[Eval].write(x))
-        ByteArray(Blake2b.hash256(bytes.value))
-      }
-    }
 
   // Dummy types for message id, sender id and transaction
   type M = ByteArray
@@ -119,7 +102,7 @@ object NetworkSim extends IOApp {
             postStateHash = postState.bytes.bytes,
           )
 
-          Block.WithId(blockBodyDigest.digest(block), block)
+          Block.WithId(block.digest, block)
         }
     }
   }
@@ -199,7 +182,7 @@ object NetworkSim extends IOApp {
           val blockSeqNumRef = Ref.unsafe(0)
           val assignBlockId  = (b: Block[M, S, T]) =>
             blockSeqNumRef.updateAndGet(_ + 1).map { seqNum =>
-              ByteArray(Blake2b.hash256(blockBodyDigest.digest(b).bytes))
+              ByteArray(Blake2b.hash256(b.digest.bytes))
             }
 
           val txSeqNumRef = Ref.unsafe(0)
